+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - 'metadata.attributions' key in ./initial-data-load/data/{dataset}/GeoData.yml configuration file as a substite for 'tileAttributions' key on the top level that is now deprecated. You can find out more about these settings in [docs](./doc/data-structures/config-files.md#geoData.yml). [#146](https://github.com/openkfw/Oscar/pull/146/files)
 - 'metadata.attributions' key in ./initial-data-load/data/{dataset}/MapLayers.yml configuration file
+- 'storeToTable' key in ./initial-data-load/data/{dataset}/GeoData.yml configuration file as a substite for 'storeToDb' and 'collectionName' key on the top level that are now deprecated. [#155](https://github.com/openkfw/Oscar/pull/155)
+- 'createTable' key in ./initial-data-load/data/{dataset}/GeoData.yml configuration file. [#155](https://github.com/openkfw/Oscar/pull/155)
 
 ### Changed
 
@@ -17,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
 - 'tileDataUrl' and 'tileAttributions' keys on the top level in ./initial-data-load/data/{country}/GeoData.yml configuration file. Use 'apiUrl' key instead of 'tileDataUrl' key and 'metadata.attributions' key instead of 'tileAttributions' key. You can find out more about these settings in [docs](./doc/data-structures/config-files.md#geoData.yml). [#146](https://github.com/openkfw/Oscar/pull/146/files)
+- 'storeToDb' and 'collectionName' key in ./initial-data-load/data/{dataset}/GeoData.yml configuration file. Use 'storeToTable' key instead. [#155](https://github.com/openkfw/Oscar/pull/155)
 
 ### Removed